[REFACTORING] tests/cypress/e2e/for-teacher_page/class_page/create_accounts_page/* (#4013)
**Description**

* `class.js::addStudents(classname, count)` creates `count` number of students with unique names to avoid collisions, returns array of student names
* `class.js::createClass()` creates a class with a unique name to avoid collisions, returns class name
* `class.js::createClassAndStudents()` a convenience function to create a class populated with 4 students,   
returns `{classname, students}`
* more specific targeting of desired class e.g.  
`cy.get(".view_class").contains(new RegExp(/^${classname}$/)).click();`

in preparation for #3993 


**How to test**
```
cd tests
npx cypress run --spec "cypress/e2e/for-teacher_page/class_page/create_accounts_page/*"
```
