+++ CONTRIBUTING.md
@@ -83,7 +83,12 @@ Do note a few things:
 * Run the `feed_dev_database.sh` script before running the tests as they something belong on certain users or classes being present in the database
 * Run pybabel before running the tests as they can also rely on exact labels
 * For the same reason, set your app to English
-
+* ensure the ADMIN_USER environment variable is set to `admin` before starting the app. e.g.     
+    ```
+    $ . ./.env/bin/activate  
+    (.env)$ export ADMIN_USER=admin  
+    (.env)$ python app.py
+    ```
 If you want to connect Cypress to the online dashboard, use:
 
 `npx cypress run --record --key <key here>`
